Moves trie into speech rule engine.
