+++ ts/rule_engine/base_rule_store.ts
@@ -25,9 +25,8 @@
 
 import {AuditoryDescription} from '../audio/auditory_description';
 import {Debugger} from '../common/debugger';
-import * as DomUtil from '../common/dom_util';
-import {Engine} from '../common/engine';
-import {Trie} from '../indexing/trie';
+// import * as DomUtil from '../common/dom_util';
+// import {Engine} from '../common/engine';
 import {Axis, AxisOrder, DynamicCstr, DynamicCstrParser} from './dynamic_cstr';
 import {Action, Precondition, SpeechRule} from './speech_rule';
 import {SpeechRuleContext} from './speech_rule_context';
@@ -41,11 +40,6 @@ export abstract class BaseRuleStore implements SpeechRuleEvaluator, SpeechRuleSt
    * Context for custom functions of this rule store.
    */
   public context: SpeechRuleContext = new SpeechRuleContext();
-
-  /**
-   * Trie for indexing speech rules in this store.
-   */
-  public trie: Trie = null;
 
   /**
    * A priority list of dynamic constraint attributes.
@@ -138,19 +132,6 @@ export abstract class BaseRuleStore implements SpeechRuleEvaluator, SpeechRuleSt
 
 
   /**
-   * Compares priority of two rules.
-   * @param rule1 The first speech rule.
-   * @param rule2 The second speech rule.
-   * @return -1, 0, 1 depending on the comparison.
-   */
-  private static priority_(rule1: SpeechRule, rule2: SpeechRule): number {
-    let priority1 = rule1.precondition.priority;
-    let priority2 = rule2.precondition.priority;
-    return priority1 === priority2 ? 0 : priority1 > priority2 ? -1 : 1;
-  }
-
-
-  /**
    * @override
    */
   constructor() {
@@ -158,34 +139,6 @@ export abstract class BaseRuleStore implements SpeechRuleEvaluator, SpeechRuleSt
       'Rule': this.defineRule,
       'Generator': this.generateRules
     };
-  }
-
-
-  /**
-   * @override
-   */
-  public lookupRule(node: Node, dynamic: DynamicCstr) {
-    if (!node ||
-        node.nodeType !== DomUtil.NodeType.ELEMENT_NODE &&
-            node.nodeType !== DomUtil.NodeType.TEXT_NODE) {
-      return null;
-    }
-    let matchingRules = this.lookupRules(node, dynamic);
-    return matchingRules.length > 0 ?
-        this.pickMostConstraint_(dynamic, matchingRules) :
-        null;
-  }
-
-
-  /**
-   * Retrieves a list of applicable rule for the given node.
-   * @param node A node.
-   * @param dynamic Additional dynamic
-   *     constraints. These are matched against properties of a rule.
-   * @return All applicable speech rules.
-   */
-  public lookupRules(node: Node, dynamic: DynamicCstr): SpeechRule[] {
-    return this.trie.lookupRules(node, dynamic.allProperties());
   }
 
 
@@ -433,47 +386,6 @@ export abstract class BaseRuleStore implements SpeechRuleEvaluator, SpeechRuleSt
     if (method) {
       method(this);
     }
-  }
-
-
-  /**
-   * Prunes the trie of the store for a given constraint.
-   * @param constraints A list of constraints.
-   */
-  public prune(constraints: string[]) {
-    let last = constraints.pop();
-    let parent = this.trie.byConstraint(constraints);
-    if (parent) {
-      parent.removeChild(last);
-    }
-  }
-
-
-  /**
-   * Picks the result of the most constraint rule by prefering those:
-   * 1) that best match the dynamic constraints.
-   * 2) with the most additional constraints.
-   * @param dynamic Dynamic constraints.
-   * @param rules An array of rules.
-   * @return The most constraint rule.
-   */
-  private pickMostConstraint_(_dynamic: DynamicCstr, rules: SpeechRule[]):
-      SpeechRule {
-    let comparator = Engine.getInstance().comparator;
-    rules.sort(function(r1, r2) {
-      return comparator.compare(r1.dynamicCstr, r2.dynamicCstr) ||
-          // When same number of dynamic constraint attributes matches for
-          // both rules, compare length of static constraints.
-          // sre.BaseRuleStore.strongQuery_(r1, r2) ||
-          BaseRuleStore.priority_(r1, r2) ||
-          r2.precondition.constraints.length -
-          r1.precondition.constraints.length ||
-          r2.precondition.rank - r1.precondition.rank;
-    });
-    Debugger.getInstance().generateOutput((() => {
-      return rules.map((x) => x.name + '(' + x.dynamicCstr.toString() + ')');
-    }).bind(this));
-    return rules[0];
   }
